Changed VGC function block to no longer use differential pressure in the interlock. Reasoning is if the operator really needs to open the valve, we have an override function. Including the differential pressure makes things too complicated.
NOTE! Also changed to using a single setpoint defined by the valve. Both
associated gauges must satisfy the valve setpoint. This is contrary to the current paradigm
of keeping the setpoints with the vacuum gauges themselves. Now we assert the device that
requires a vacuum pressure should check for itself if the interlock is satisfied.

Added a hysteresis boundary to the vacuum setpoint for valves. Set it up so the operator can
select a vacuum setpoint and hyst. boundary, and the two will move proportionally if the vac sp
is moved later. The hyst boundary can be moved independently of the vac sp.

We plan to show the vac sp in the normal operator screen, and the vac sp and hyst in an expert panel.


git-svn-id: file:///reg/g/pcds/plc-common/plcRepo/libraries/LCLSVacuumLib@765 7fb2447c-b5a5-4d0a-8d33-1ac561a9f937